Completed the combinational and sequential logic associated with read/write operationsmemory.sv (#3)
NOTE: The issue in gtkwave pertains to the fact that at initialization, read_enable_i does not immediately start with 1 at t = 0 but after approximately 2 ps. This is problematic, as fetch.sv's program counter manages to increment itself and update the instruction at the rising clock edge each time. While it may seem correct in gtkwave, data_out stores the wrong output at each read when an instruction is fetched. Please observe the wave for further detail and the video on expected output for PD1 on EClass.